ctfs/InsomniHack2/2024/Quals/web/InsoBank/api/requirements.txt
Flask==3.0.0
mysql-connector-python==8.0.27
flask_jwt_extended
flask-cors
psycopg2
pyopenssl